+++ changelog.md
@@ -52,6 +52,7 @@ Codion Change Log
 - TabbedDetailLayout, keyboard shortcuts added for expanding the detail panel.
 - TabbedDetailLayout, now shows the detail panel if hidden when activated.
 - EntityTablePanel.createControls() bug fixed, now only relies on configuration.
+- EntityTablePanel.Config bug fixed, now initializes includeSummaryPanel to false if no summary models are available.
 
 ## 0.17.43
 ### is.codion.common.core